Improve magic ship
Prefix sums can be used for the weather cycle.
I was pulling my hair out
because the `pdx[i] = pdx[i-1]` update was inside if-else.
It doesn't work this way since the horizontal state has to propagate
even when the wind direction is 'U' or 'D'.
Be careful about coordinate prefix sums in the future.